prosto:

+++ includes/file.h
@@ -6,13 +6,15 @@
 /*   By: tjuana <tjuana@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/17 15:07:53 by tjuana            #+#    #+#             */
-/*   Updated: 2020/01/18 20:22:47 by tjuana           ###   ########.fr       */
+/*   Updated: 2020/01/20 19:40:01 by tjuana           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "wolf3d.h"
-# define VER_HEIGHT (int)33
-# define VER_WIDTH (int)54
+#include <sys/types.h>
+#include <sys/stat.h>
+# define VER_HEIGHT 33
+# define VER_WIDTH 54
 
 
 typedef struct			s_file
